Controller Montado

+++ src/Backend/MinhaAgendaDeConsultas.Domain/Entidades/AgendamentoConsultas.cs
@@ -4,8 +4,8 @@ namespace MinhaAgendaDeConsultas.Domain.Entidades
 {
     public class AgendamentoConsultas : EntidadeBase
     {
-        public int PacienteId { get; set; }
-        public int MedicoId { get; set; }       
+        public long PacienteId { get; set; }
+        public long MedicoId { get; set; }       
         public DateTime DataInclusao { get; set; }
         public DateTime DataHoraInicio { get; set; }
         public DateTime DataHoraFim { get; set; }
