+++ src/Shared/MinhaAgendaDeConsultas.Communication/Resposta/Agendamento/ResponseConsultaAgendamentos.cs
@@ -9,8 +9,8 @@ namespace MinhaAgendaDeConsultas.Communication.Resposta.Agendamento
 {
     public  class ResponseConsultaAgendamentos
     {
-        public int PacienteId { get; set; }
-        public int MedicoId { get; set; }  
+        public long PacienteId { get; set; }
+        public long MedicoId { get; set; }  
         public DateTime DataInclusao { get; set; }
         public DateTime DataHoraInicio { get; set; }
         public DateTime DataHoraFim { get; set; }
